+++ com.devsquad10.product/src/main/java/com/devsquad10/product/domain/repository/ProductQuerydslRepository.java
@@ -1,0 +1,9 @@
+package com.devsquad10.product.domain.repository;
+
+import org.springframework.data.domain.Page;
+
+import com.devsquad10.product.domain.model.Product;
+
+public interface ProductQuerydslRepository {
+	Page<Product> findAll(String q, String category, int page, int size, String sort, String order);
+}
