+++ com.devsquad10.product/src/main/java/com/devsquad10/product/domain/repository/ProductRepository.java
@@ -3,8 +3,6 @@ package com.devsquad10.product.domain.repository;
 import java.util.Optional;
 import java.util.UUID;
 
-import org.springframework.data.domain.Page;
-
 import com.devsquad10.product.domain.model.Product;
 
 public interface ProductRepository {
@@ -12,8 +10,6 @@ public interface ProductRepository {
 	Optional<Product> findByIdAndDeletedAtIsNull(UUID id);
 
 	Product save(Product product);
-
-	Page<Product> findAll(String q, String category, int page, int size, String sort, String order);
 
 	int decreaseStock(UUID productId, int orderQuantity);
 }
